Abstract data source default button component
